[refactor] 디렉토리 수정 및 코드 리팩토링 #83
